feat: add an UI option to ignore conceal warnings
In https://github.com/epwalsh/obsidian.nvim/issues/286#issuecomment-2048334358
it was suggested that a new config could be added to specifically not
warn if the conceal level is 0 and the UI is enabled.

This commit does this and guard against the conceal warning when the
option is set.

cc https://github.com/epwalsh/obsidian.nvim/issues/492